requirements.txt
# Telegram Bot
python-telegram-bot==21.5

# OCR
pytesseract==0.3.13
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

# LLM (Ollama API)
httpx==0.27.2

# Google Calendar
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Utils
python-dotenv==1.0.1
